Disable oo7 default-features in workspace.dependencies
This is to stop oo7/native_crypto feature being enabled when using
oo7/openssl_crypto feature on the server side (or the opposite).

This also resolves the following error when setting oo7
`default-features = false` in workspace members.
"`default-features` is ignored for oo7, since `default-features` was not
specified for `workspace.dependencies.oo7`, this could become a hard error
in the future".

Signed-off-by: Dhanuka Warusadura <dhanuka@gnome.org>