portal: Dont ask for app tokens with schema
If gnome-keyring creates a secret without a schema, due to
https://github.com/bilelmoussaoui/oo7/issues/233, the search would
return no values and a new secret would be stored by the app. This would
result in the app getting the wrong secret when using the portal.

It could be argued that https://github.com/bilelmoussaoui/oo7/issues/233
is the real issue at place, but the generic schema is the default schema
and assumed if no schema is set and it does not make sense to search for
secrets having it.