client: key: Remove impl From<&key> for Value
These methods require allocating and taking a reference in a From is not
very idiomatic, instead we use the inner Vec and the implementation of
From<Vec<u8>> of zvariant::Array.

Note that since Key implements Drop, we cannot just use self.key and we
need to take the value out.